+++ packages/shared/src/index.ts
@@ -16,3 +16,4 @@
 export * from './types.js';
 export * from './constants.js';
 export * from './utils.js';
+export * from './approvals.js';